adminservice/adminapi/serializers.py
from rest_framework import serializers
from .models import Rooms

class RoomSerializer(serializers.ModelSerializer):
    class Meta:
        model = Rooms
        fields = '__all__'  # Wszystkie pola modelu
